fix(kitchen): move dishUpdateBody out of use-server module
Next.js requires every export from a Server Actions file to be async. Keep the full-body Dish.update helpers in a plain module and import them from the actions.
